break down the admitters tests
