+++ pkg/virt-api/webhooks/validating-webhook/admitters/vmi-update-admitter_test.go
@@ -1,0 +1,97 @@
+/*
+ * This file is part of the KubeVirt project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * Copyright 2018 Red Hat, Inc.
+ *
+ */
+
+package admitters
+
+import (
+	"encoding/json"
+
+	. "github.com/onsi/ginkgo"
+	"github.com/onsi/ginkgo/extensions/table"
+	. "github.com/onsi/gomega"
+	"k8s.io/api/admission/v1beta1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+
+	v1 "kubevirt.io/kubevirt/pkg/api/v1"
+	"kubevirt.io/kubevirt/pkg/virt-api/webhooks"
+)
+
+var _ = Describe("Validating VMIUpdate Admitter", func() {
+	vmiUpdateAdmitter := &VMIUpdateAdmitter{}
+
+	table.DescribeTable("should reject documents containing unknown or missing fields for", func(data string, validationResult string, gvr metav1.GroupVersionResource, review func(ar *v1beta1.AdmissionReview) *v1beta1.AdmissionResponse) {
+		input := map[string]interface{}{}
+		json.Unmarshal([]byte(data), &input)
+
+		ar := &v1beta1.AdmissionReview{
+			Request: &v1beta1.AdmissionRequest{
+				Resource: gvr,
+				Object: runtime.RawExtension{
+					Raw: []byte(data),
+				},
+			},
+		}
+		resp := review(ar)
+		Expect(resp.Allowed).To(BeFalse())
+		Expect(resp.Result.Message).To(Equal(validationResult))
+	},
+		table.Entry("VirtualMachineInstance update",
+			`{"very": "unknown", "spec": { "extremely": "unknown" }}`,
+			`.very in body is a forbidden property, spec.extremely in body is a forbidden property, spec.domain in body is required`,
+			webhooks.VirtualMachineInstanceGroupVersionResource,
+			vmiUpdateAdmitter.Admit,
+		),
+	)
+
+	It("should reject valid VirtualMachineInstance spec on update", func() {
+		vmi := v1.NewMinimalVMI("testvmi")
+
+		updateVmi := vmi.DeepCopy()
+		updateVmi.Spec.Domain.Devices.Disks = append(vmi.Spec.Domain.Devices.Disks, v1.Disk{
+			Name: "testdisk",
+		})
+		updateVmi.Spec.Volumes = append(vmi.Spec.Volumes, v1.Volume{
+			Name: "testdisk",
+			VolumeSource: v1.VolumeSource{
+				ContainerDisk: &v1.ContainerDiskSource{},
+			},
+		})
+		newVMIBytes, _ := json.Marshal(&updateVmi)
+		oldVMIBytes, _ := json.Marshal(&vmi)
+
+		ar := &v1beta1.AdmissionReview{
+			Request: &v1beta1.AdmissionRequest{
+				Resource: webhooks.VirtualMachineInstanceGroupVersionResource,
+				Object: runtime.RawExtension{
+					Raw: newVMIBytes,
+				},
+				OldObject: runtime.RawExtension{
+					Raw: oldVMIBytes,
+				},
+				Operation: v1beta1.Update,
+			},
+		}
+
+		resp := vmiUpdateAdmitter.Admit(ar)
+		Expect(resp.Allowed).To(Equal(false))
+		Expect(len(resp.Result.Details.Causes)).To(Equal(1))
+		Expect(resp.Result.Details.Causes[0].Message).To(Equal("update of VMI object is restricted"))
+	})
+})
